requirements.txt
msal
pyyaml
tqdm